Refactor table.dart using part files to make it more manageable.
